+++ explorateurs_autonomes/src/data/map/mobile/MobileElement.java
@@ -16,7 +16,7 @@ import data.map.geometry.Position;
  * */
 
 public class MobileElement extends GraphicElement{
-	private float pace;
+	private int pace;
 	private String name;
 	private int health;
 	private int strength;
@@ -27,7 +27,7 @@ public class MobileElement extends GraphicElement{
 	public static final int MONSTER_HEALTH = 250;
 	
 	
-	public MobileElement(Position position, Dimension dimension, float pace, String name, 
+	public MobileElement(Position position, Dimension dimension, int pace, String name, 
 			int health, int strength, CircularSector vision, Sound sound) {
 		super(dimension, position);
 		this.pace = pace;
@@ -39,12 +39,12 @@ public class MobileElement extends GraphicElement{
 	}
 	
 	
-	public MobileElement(Position position, Dimension dimension, float pace, String name, 
+	public MobileElement(Position position, Dimension dimension, int pace, String name, 
 			int strength, CircularSector vision, Sound sound) {
 		this(position, dimension, pace, name, CHARACTER_HEALTH, strength, vision, sound);
 	}
 	
-	public MobileElement(Position position, Dimension dimension, float pace, String name,
+	public MobileElement(Position position, Dimension dimension, int pace, String name,
 			int health, int strength, Sound sound) {
 		this(position, dimension, pace, name, health, strength, new CircularSector(5,6,7,8),
 				sound);
@@ -61,12 +61,12 @@ public class MobileElement extends GraphicElement{
 	}
 
 
-	public float getPace() {
+	public int getPace() {
 		return pace;
 	}
 	
 	
-	public void setPace(float pace) {
+	public void setPace(int pace) {
 		this.pace = pace;
 	}
